otoritas di logistic

+++ trunk/application/views/master/kategori.php
@@ -93,10 +93,13 @@
 		
 		actionbutton = function(value, row, index){
 			var col;
-			//if (row.kd_fakultas != null) {
+			<?if($this->mdl_auth->CekAkses(array('menu_id'=>6, 'policy'=>'EDIT'))){?>
 				col = '<a href="#" onclick="editData(\''+row. 	id_kategori+'\');" class="easyui-linkbutton" iconCls="icon-edit" plain="false">Edit</a>';
+			<?}?>
+
+			<?if($this->mdl_auth->CekAkses(array('menu_id'=>6, 'policy'=>'DELETE'))){?>
 				col += '&nbsp;&nbsp;|&nbsp;&nbsp;<a href="#" onclick="deleteData(\''+row. 	id_kategori+'\');" class="easyui-linkbutton" iconCls="icon-edit" plain="false">Delete</a>';
-			//}
+			<?}?>
 			return col;
 		}
 		
@@ -111,6 +114,7 @@
 			var pager = $('#dg').datagrid().datagrid('getPager');	// get the pager of datagrid
 			pager.pagination({
 				buttons:[
+				<?if($this->mdl_auth->CekAkses(array('menu_id'=>6, 'policy'=>'ADD'))){?>
 					{
 						iconCls:'icon-add',
 						text:'Tambah Data',
@@ -118,6 +122,7 @@
 							newData();
 						}
 					}
+				<?}?>	
 				]
 			});			
 		});
