+++ trunk/application/views/transfer/transfer.php
@@ -107,13 +107,15 @@
 
 		actionbutton = function(value, row, index){
 			var col='';
-			
+				<?if($this->mdl_auth->CekAkses(array('menu_id'=>43, 'policy'=>'DETAIL'))){?>
 					col += '<a href="#" onclick="detailData(\''+row.id_transfer+'\');" class="easyui-linkbutton" iconCls="icon-edit" plain="false">Detail</a>';
-			
+				<?}?>
+				<?if($this->mdl_auth->CekAkses(array('menu_id'=>43, 'policy'=>'APPROVE'))){?>
 					col += '&nbsp;&nbsp;|&nbsp;&nbsp;<a href="#" onclick="doneData(\''+row.id_transfer+'\');" class="easyui-linkbutton" iconCls="icon-edit" plain="false">Done</a>';
-		
+				<?}?>
+				<?if($this->mdl_auth->CekAkses(array('menu_id'=>43, 'policy'=>'DELETE'))){?>
 					col += '&nbsp;&nbsp;|&nbsp;&nbsp;<a href="#" onclick="deleteData(\''+row.id_transfer+'\');" class="easyui-linkbutton" iconCls="icon-edit" plain="false">Delete</a>';
-			
+				<?}?>
 			return col;
 		}
 		
@@ -144,7 +146,7 @@
 			var pager = $('#dg').datagrid().datagrid('getPager');	// get the pager of datagrid
 			pager.pagination({
 				buttons:[
-				<?if($this->mdl_auth->CekAkses(array('menu_id'=>10, 'policy'=>'ADD'))){?>
+				<?if($this->mdl_auth->CekAkses(array('menu_id'=>43, 'policy'=>'ADD'))){?>
 					{
 						iconCls:'icon-add',
 						text:'Tambah Data',
